test_no_adk/requirements.txt
fastapi
uvicorn
google-genai
websockets
numpy
soxr
python-multipart
twilio
python-dotenv
pyaudio
